Modifying  the price TextView in the general ListView(?) and modyfing the look of the purchase details
The textView in the general ListView used for prices has been modified (subject to change). Some changes have been modified in order to make some objects fit and look better in some activities, specially in the pruchase details one, which has been change in order to make the interface look better when you change the orientation to landscape mode.