Chaning the look of the CardView in properties + more colors and project settings
The previous look of the CardView template for the categories layout has been changed (both in its design and implementation in the categories Acitivity). Also, new colors have been added to the project in the colors file. Also, in order to work with some functions of the projects, the gradle of the project has been changed